Region||voxelPack

+++ GLFW_tutorial/source/Game/LightSolver.h
@@ -18,8 +18,8 @@ public:
 	}
 	void add(const glm::ivec3& pos);
 	inline void add(int x, int y, int z) { add(glm::ivec3(x, y, z)); }
-	void add(const glm::ivec3& pos, int emission);
-	inline void add(int x, int y, int z, int emission) { add(glm::ivec3(x, y, z), emission); }
+	void add(const glm::ivec3& pos, byte emission);
+	inline void add(int x, int y, int z, byte emission) { add(glm::ivec3(x, y, z), emission); }
 	void remove(const glm::ivec3& pos);
 	inline void remove(int x, int y, int z) { remove(glm::ivec3(x, y, z)); }
 	void solve();
